+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # OpenApps
 
+## 0.0.7
+
+- Fixed authentication functions pointing to old `it.teithe` servers.
+- Updated features that depend on the official Apps platform for correct information.
+
 ## 0.0.6
 
 - Updated to Angular 12.1.1